requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
drf_spectacular
django-filter
django-redis
Pillow